+++ public/icons/name.d.ts
@@ -56,6 +56,7 @@
 	| "flame"
 	| "gas_xl"
 	| "gas"
+	| "gear_slim"
 	| "gear"
 	| "globe-b"
 	| "globe"
