+++ ui/snippets/topBar/settings/Settings.tsx
@@ -1,0 +1,33 @@
+import { IconButton, Popover, PopoverBody, PopoverContent, PopoverTrigger, useDisclosure } from '@chakra-ui/react';
+import React from 'react';
+
+import IconSvg from 'ui/shared/IconSvg';
+
+import SettingsColorTheme from './SettingsColorTheme';
+
+const Settings = () => {
+  const { isOpen, onToggle, onClose } = useDisclosure();
+
+  return (
+    <Popover placement="bottom-start" trigger="click" isOpen={ isOpen } onClose={ onClose }>
+      <PopoverTrigger>
+        <IconButton
+          variant="simple"
+          colorScheme="blue"
+          aria-label="color mode switch"
+          icon={ <IconSvg name="gear_slim" boxSize={ 5 }/> }
+          p="1px"
+          boxSize={ 5 }
+          onClick={ onToggle }
+        />
+      </PopoverTrigger>
+      <PopoverContent overflowY="hidden" w="auto" fontSize="sm">
+        <PopoverBody boxShadow="2xl" p={ 4 }>
+          <SettingsColorTheme/>
+        </PopoverBody>
+      </PopoverContent>
+    </Popover>
+  );
+};
+
+export default React.memo(Settings);
